refactor: store punishments as epoch seconds
Longs aren't supported by the variable format system; sub-second
precision isn't needed for mutes, bans, black marks or reports.
Switch them to epoch seconds (Int), with PERMANENT as Int.MAX_VALUE.
Chat message evidence keeps millisecond precision.

+++ engine/src/main/kotlin/world/gregs/voidps/engine/data/AbuseReport.kt
@@ -8,7 +8,7 @@ package world.gregs.voidps.engine.data
  * @param ruleName Readable name of the rule broken
  * @param mute Whether a moderator requested the accused be muted
  * @param suggestion Additional text submitted with the report
- * @param time Epoch millisecond timestamp the report was received
+ * @param time Epoch second timestamp the report was received
  * @param evidence Recent chat messages sent by the accused player
  */
 data class AbuseReport(
@@ -18,6 +18,6 @@ data class AbuseReport(
     val ruleName: String,
     val mute: Boolean,
     val suggestion: String,
-    val time: Long,
+    val time: Int,
     val evidence: List<String>,
 )
